requirements/requirement_32.txt
web.py>=0.30
requests>=2.11.0
lxml-3.6.4-cp27-cp27m-win32.whl